Split test_parse.py into strict and unsafe halves
The parse_string refactor made the split obvious. TestParseStrict
covers the default strict=True path plus public entry-point tests
(including parse_string_unsafe's deprecation shim). TestParseUnsafe
covers strict=False.

314 passing, 2 skipped. Up from 303 before this branch.
